extract the process of `rules` subcommand into rules.rs
