+++ examples/dlint/main.rs
@@ -11,10 +11,9 @@ use deno_lint::ast_parser::{get_default_es_config, get_default_ts_config};
 use deno_lint::diagnostic::LintDiagnostic;
 use deno_lint::diagnostic::Range;
 use deno_lint::linter::LinterBuilder;
-use deno_lint::rules::{get_all_rules, get_recommended_rules};
+use deno_lint::rules::get_recommended_rules;
 use log::debug;
 use rayon::prelude::*;
-use serde::Serialize;
 use std::collections::BTreeMap;
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicUsize, Ordering};
@@ -24,6 +23,7 @@ use swc_ecmascript::parser::{EsConfig, Syntax, TsConfig};
 
 mod config;
 mod js;
+mod rules;
 
 fn create_cli_app<'a, 'b>() -> App<'a, 'b> {
   App::new("dlint")
@@ -263,86 +263,6 @@ fn determine_syntax(path: &Path) -> Syntax {
   }
 }
 
-#[derive(Clone, Copy, Serialize)]
-struct Rule {
-  code: &'static str,
-  docs: &'static str,
-  tags: &'static [&'static str],
-}
-
-fn get_all_rules_metadata() -> Vec<Rule> {
-  get_all_rules()
-    .into_iter()
-    .map(|rule| Rule {
-      code: rule.code(),
-      docs: rule.docs(),
-      tags: rule.tags(),
-    })
-    .collect()
-}
-
-trait RuleFormatter {
-  fn format(rules: &mut [Rule]) -> Result<String, &'static str>;
-}
-
-enum JsonFormatter {}
-enum PrettyFormatter {}
-
-impl RuleFormatter for JsonFormatter {
-  fn format(rules: &mut [Rule]) -> Result<String, &'static str> {
-    if rules.is_empty() {
-      return Err("Rule not found!");
-    }
-    serde_json::to_string_pretty(rules).map_err(|_| "failed to format!")
-  }
-}
-
-impl RuleFormatter for PrettyFormatter {
-  fn format(rules: &mut [Rule]) -> Result<String, &'static str> {
-    if rules.is_empty() {
-      return Err("Rule not found!");
-    }
-
-    if rules.len() == 1 {
-      let rule = &rules[0];
-      let docs = if rule.docs.is_empty() {
-        "documentation not available"
-      } else {
-        rule.docs
-      };
-      return Ok(format!("- {code}\n\n{docs}", code = rule.code, docs = docs));
-    }
-
-    rules.sort_by_key(|r| r.code);
-    let mut list = Vec::with_capacity(1 + rules.len());
-    list.push("Available rules (trailing ✔️ mark indicates it is included in the recommended rule set):".to_string());
-    list.extend(rules.iter().map(|r| {
-      let mut s = format!(" - {}", r.code);
-      if r.tags.contains(&"recommended") {
-        s += " ✔️";
-      }
-      s
-    }));
-    Ok(list.join("\n"))
-  }
-}
-
-fn print_rules<F: RuleFormatter>(mut rules: Vec<Rule>) {
-  match F::format(&mut rules) {
-    Err(e) => {
-      eprintln!("{}", e);
-      std::process::exit(1);
-    }
-    Ok(text) => {
-      println!("{}", text);
-    }
-  }
-}
-
-fn filter_rules(rules: Vec<Rule>, rule_name: &str) -> Vec<Rule> {
-  rules.into_iter().filter(|r| r.code == rule_name).collect()
-}
-
 fn main() -> Result<(), AnyError> {
   env_logger::init();
 
@@ -383,16 +303,15 @@ fn main() -> Result<(), AnyError> {
       )?;
     }
     ("rules", Some(rules_matches)) => {
-      let json = rules_matches.is_present("json");
       let rules = if let Some(rule_name) = rules_matches.value_of("RULE_NAME") {
-        filter_rules(get_all_rules_metadata(), rule_name)
+        rules::get_specific_rule_metadata(rule_name)
       } else {
-        get_all_rules_metadata()
+        rules::get_all_rules_metadata()
       };
-      if json {
-        print_rules::<JsonFormatter>(rules);
+      if rules_matches.is_present("json") {
+        rules::print_rules::<rules::JsonFormatter>(rules);
       } else {
-        print_rules::<PrettyFormatter>(rules);
+        rules::print_rules::<rules::PrettyFormatter>(rules);
       }
     }
     _ => unreachable!(),
